docs(all): included docstrings for everything! (#49)
* chore(all): configured flake8 and remediated

* chore(all): migrated from np.float64 to python float

* chore(all): introduced flake8 and remediated issues

* chore(deps-dev): bump hypothesis from 5.41.2 to 5.41.3 (#45)

* chore(deps-dev): bump hypothesis from 5.41.2 to 5.41.3

Bumps [hypothesis](https://github.com/HypothesisWorks/hypothesis) from 5.41.2 to 5.41.3.
- [Release notes](https://github.com/HypothesisWorks/hypothesis/releases)
- [Commits](https://github.com/HypothesisWorks/hypothesis/compare/hypothesis-python-5.41.2...hypothesis-python-5.41.3)


* Updated poetry.lock to reflect new dependency versions

* Updated poetry.lock to reflect new dependency versions


* chore(docs): Wrote docstrings for everything!

* Updated circle to use custom image.

* chore(deps-dev): bump darglint from 1.5.6 to 1.5.8 (#47)

* Updated circle to use custom image.

* chore(deps-dev): bump darglint from 1.5.6 to 1.5.8

Bumps [darglint](https://github.com/raddessi/SublimeLinter-contrib-darglint) from 1.5.6 to 1.5.8.
- [Release notes](https://github.com/raddessi/SublimeLinter-contrib-darglint/releases)
- [Commits](https://github.com/raddessi/SublimeLinter-contrib-darglint/commits)



* chore(deps): bump importlib-metadata from 2.0.0 to 3.1.0 (#48)

* chore(deps): bump importlib-metadata from 2.0.0 to 3.1.0

Bumps [importlib-metadata](https://github.com/python/importlib_metadata) from 2.0.0 to 3.1.0.
- [Release notes](https://github.com/python/importlib_metadata/releases)
- [Changelog](https://github.com/python/importlib_metadata/blob/master/CHANGES.rst)
- [Commits](https://github.com/python/importlib_metadata/compare/v2.0.0...v3.1.0)


* Updated poetry.lock to reflect new dependency versions


* build: removed autodoc_typehints whilst build is failing

+++ tests/test_shape.py
@@ -1,6 +1,4 @@
-import numpy as np
-
-from pysketcher import Arrow, Axis, Force, Line, Point, Shape, Style
+from pysketcher import Shape, Style
 from tests.utils import given_inferred
 
 
@@ -12,7 +10,7 @@ class TestShape:
         assert shape.style == style
 
     @given_inferred
-    def test_set_line_width(self, shape: Shape, line_width: np.float64):
+    def test_set_line_width(self, shape: Shape, line_width: float):
         new_shape = shape.set_line_width(line_width)
         assert new_shape == shape
         assert new_shape.style.line_width == line_width
@@ -48,7 +46,7 @@ class TestShape:
         assert new_shape.style.arrow == arrow
 
     @given_inferred
-    def test_set_shadow(self, shape: Shape, shadow: np.float64):
+    def test_set_shadow(self, shape: Shape, shadow: float):
         new_shape = shape.set_shadow(shadow)
         assert new_shape == shape
         assert new_shape.style.shadow == shadow
